# lecture-11-16/student-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student API",
        "description": "API for students with authentication, posts, likes, comments, shares, and friends.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://courses.xrow.asia/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/friend-request/{user}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Send friend request",
                "operationId": "8ff762f454662f502d1c439b3cb30743",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Request sent"
                    },
                    "409": {
                        "description": "Already sent or friends"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Reject friend request",
                "operationId": "5c663c41d14817727b99477dbd7bbc16",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request rejected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friend-request/{user}/accept": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Accept friend request",
                "operationId": "a811b75b5229d79810ddbd9c07080b35",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request accepted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friends/{user}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove a friend",
                "operationId": "4a675e9e233326c886f11f1ea6882fd6",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "List friends",
                "operationId": "cb1684b02952c3db39d7240387c2cac5",
                "responses": {
                    "200": {
                        "description": "Paginated friends list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friend-requests/pending": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Pending received requests",
                "operationId": "3025217061176054e003cbf688ea7402",
                "responses": {
                    "200": {
                        "description": "Paginated pending requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friend-requests/sent": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Sent requests",
                "operationId": "1933be0419cf01161a134d1d18d86e79",
                "responses": {
                    "200": {
                        "description": "Paginated sent requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/toggle-like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Toggle like",
                "operationId": "26a8bd18e9bfeb9c5f863f7cd6f30c90",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like a post",
                "operationId": "466e3ae39106d7c67096622459c5e09b",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post liked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Unlike a post",
                "operationId": "72244834f1ef3de66dd7337fc79ecaf7",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like removed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}